+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJA0G6E45VSCAFPESDA5EWDT"
+        "01KJAAPMCA5YVSB9T510KGGHSB"
     )
     r"""The unique identifier of the data source"""