+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJA0G6E6DT0QQWTAND9J9WXZ"
+        "01KJAAPMCCHQ21V7A5TXCC8F7Z"
     )
     r"""The unique identifier of the data source"""