+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJA0G6E9A9WM423GJQ74TJRT"
+        "01KJAAPMCJMVF9QG6K13CEDTNT"
     )
     r"""The unique identifier of the data source"""