+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -259,7 +259,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KJA0G69E9QXSNM5EBYX5FE1T"
+    id: Optional[str] = "01KJAAPM8MPW4XKX0MJP0ECGH9"
 
     description: Optional[str] = None
 
@@ -1214,7 +1214,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJA0G696P1HJ89KWZTEKAQBF"
+        "tool_01KJAAPM8ETXVSYJB9K84JWQY6"
     )
 
     display_name: Optional[str] = None
@@ -1345,7 +1345,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KJA0G695A8PWPA3MJF9B811D"
+    id: Optional[str] = "01KJAAPM8D6P7ACX3RS8XGG1V5"
 
     description: Optional[str] = None
 
@@ -1476,7 +1476,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJA0G694FT6JGT452PBQRZ7H"
+        "tool_01KJAAPM8BQ78H89YCCX7D4MH5"
     )
 
     display_name: Optional[str] = None
@@ -1777,7 +1777,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJA0G692Z205E85NNN59ZZYS"
+        "tool_01KJAAPM88AZTMNFF44JTSH49P"
     )
 
     display_name: Optional[str] = None
@@ -1972,7 +1972,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJA0G691RFWH9JVVYDBB9HYV"
+        "tool_01KJAAPM86Z8GHM4QRGSE9Y9XH"
     )
 
     display_name: Optional[str] = None
@@ -2171,7 +2171,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJA0G68Z2YMD471K477YSDA9"
+        "tool_01KJAAPM84HBD5BCHTBTDZXAYF"
     )
 
     display_name: Optional[str] = None